review: address comments on #9410
- Inline `classifyCloseEvent` into `useMarimoKernelConnection.tsx` and
  delete `close-handler.ts`; co-locate the test alongside the consumer.
- Use `logNever` in the default branch of the close-reason switch so a
  new MARIMO_* reason added server-side surfaces a console warning
  instead of falling silently into the retry path. Move the retry/
  give-up logic after the switch.
- `reconnect()`: early-return on `CONNECTING` as well as `OPEN` so a
  double-click can't fire a duplicate request.
- `reconnect()`: reset `shouldTryReconnecting.current = false` on
  health-probe failure so auto-reconnect doesn't keep firing against
  an unreachable runtime.
