+++ frontend/src/core/websocket/useMarimoKernelConnection.tsx
@@ -11,7 +11,6 @@ import type {
   NotificationMessageData,
   NotificationPayload,
 } from "@/core/kernel/messages";
-import { classifyCloseEvent } from "@/core/websocket/close-handler";
 import {
   MAX_RETRIES,
   useConnectionTransport,
@@ -73,9 +72,91 @@ import { useStorageActions } from "../storage/state";
 import { useVariablesActions } from "../variables/state";
 import type { VariableName } from "../variables/types";
 import { isWasm } from "../wasm/utils";
-import { WebSocketClosedReason, WebSocketState } from "./types";
+import { type ConnectionStatus, WebSocketClosedReason, WebSocketState } from "./types";
 
 const SUPPORTS_LAZY_KERNELS = true;
+
+export type CloseDecision =
+  | { kind: "terminal"; status: ConnectionStatus; closeTransport: boolean }
+  | { kind: "gave-up"; status: ConnectionStatus }
+  | { kind: "retry"; status: ConnectionStatus };
+
+export function classifyCloseEvent(
+  event: { reason?: string },
+  context: { retryCount: number; maxRetries: number },
+): CloseDecision {
+  switch (event.reason) {
+    case "MARIMO_ALREADY_CONNECTED":
+      return {
+        kind: "terminal",
+        status: {
+          state: WebSocketState.CLOSED,
+          code: WebSocketClosedReason.ALREADY_RUNNING,
+          reason: "another browser tab is already connected to the kernel",
+          canTakeover: true,
+        },
+        closeTransport: true,
+      };
+    case "MARIMO_WRONG_KERNEL_ID":
+    case "MARIMO_NO_FILE_KEY":
+    case "MARIMO_NO_SESSION_ID":
+    case "MARIMO_NO_SESSION":
+    case "MARIMO_SHUTDOWN":
+      return {
+        kind: "terminal",
+        status: {
+          state: WebSocketState.CLOSED,
+          code: WebSocketClosedReason.KERNEL_DISCONNECTED,
+          reason: "kernel not found",
+        },
+        closeTransport: true,
+      };
+    case "MARIMO_MALFORMED_QUERY":
+      return {
+        kind: "terminal",
+        status: {
+          state: WebSocketState.CLOSED,
+          code: WebSocketClosedReason.MALFORMED_QUERY,
+          reason:
+            "the kernel did not recognize a request; please file a bug with marimo",
+        },
+        closeTransport: false,
+      };
+    case "MARIMO_KERNEL_STARTUP_ERROR":
+      return {
+        kind: "terminal",
+        status: {
+          state: WebSocketState.CLOSED,
+          code: WebSocketClosedReason.KERNEL_STARTUP_ERROR,
+          reason: "Failed to start kernel sandbox",
+        },
+        closeTransport: true,
+      };
+    default:
+      // Empty/undefined reasons are normal transient closes. Anything else is
+      // an unknown server reason; warn so a new MARIMO_* reason doesn't fall
+      // silently into the retry path.
+      if (event.reason) {
+        logNever(event.reason as never);
+      }
+  }
+  // partysocket stops retrying silently once `maxRetries` is hit; surface
+  // CLOSED so callers can detect the give-up.
+  if (context.retryCount >= context.maxRetries) {
+    return {
+      kind: "gave-up",
+      status: {
+        state: WebSocketState.CLOSED,
+        code: WebSocketClosedReason.KERNEL_DISCONNECTED,
+        reason: "kernel not found",
+      },
+    };
+  }
+  return {
+    kind: "retry",
+    status: { state: WebSocketState.CONNECTING },
+  };
+}
 
 function getExistingCells(): CellData[] | undefined {
   if (!SUPPORTS_LAZY_KERNELS) {
@@ -347,13 +428,17 @@ export function useMarimoKernelConnection(opts: {
   // Manual reconnect. Probes /health first to fail fast when the runtime
   // is unreachable, instead of waiting on partysocket's retry budget.
   const reconnect = async () => {
-    if (ws.readyState === WebSocket.OPEN) {
+    if (
+      ws.readyState === WebSocket.OPEN ||
+      ws.readyState === WebSocket.CONNECTING
+    ) {
       return;
     }
     shouldTryReconnecting.current = true;
     setConnection({ state: WebSocketState.CONNECTING });
     const healthy = await runtimeManager.isHealthy();
     if (!healthy) {
+      shouldTryReconnecting.current = false;
       setConnection({
         state: WebSocketState.CLOSED,
         code: WebSocketClosedReason.KERNEL_DISCONNECTED,
